backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
pydantic-settings==2.4.0
bcrypt==4.2.0
itsdangerous==2.2.0
python-multipart==0.0.9
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
